fix : 게시글 좋아요 방식변경
필드에 저장하지않고 post like 테이블에서 카운트해옴

+++ src/main/java/com/potato/petpotatocommunity/service/PostLikeService.java
@@ -2,6 +2,6 @@ package com.potato.petpotatocommunity.service;
 
 public interface PostLikeService {
     boolean isLiked(Long postId, Long userId);
-    int getLikeCount(Long postId);
+    long getLikeCount(Long postId);
     boolean toggleLike(Long postId, Long userId);
 }
